♻️ move custom serializer fields to serperate file
